+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+
+## [Unreleased]
+
+### Changed
+- support package translations
+- support multiple categories
+
+### Fixed
+- store function binding
+
 ## 1.3.7
 
 ### Changed